pyproject.toml
[project]
name = "local-agent-substrate"
version = "0.1.0"
description = "Reusable local AI-agent substrate with automated prompt chaining."
readme = "README.md"
license = { text = "GPL-3.0-or-later" }
requires-python = ">=3.12,<3.13"
dependencies = [
  "langchain>=0.3.0",
  "langgraph>=0.2.0",
  "langchain-openai>=0.2.0",
  "langchain-anthropic>=0.2.0",
  "langchain-ollama>=0.2.0",
  "mkdocs>=1.6.0",
  "mkdocs-material>=9.6.0",
  "fastapi>=0.116.0",
  "uvicorn>=0.30.0",
  "jinja2>=3.1.0",
  "python-multipart>=0.0.20",
  "pyyaml>=6.0.2",
  "python-dotenv>=1.0.1",
  "rich>=13.9.4",
]

[project.optional-dependencies]
dev = [
  "pre-commit>=4.0.0",
  "ruff>=0.11.0",
]

[tool.uv]
package = false
